+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 * Moved configuration options 'encryption.aes', 'encryption.arc4',
   'encryption.filter_map', 'encryption.sub_filter.map', 'filter.map',
   'image_loader' and 'task.map' to the document specific configuration object
+* [HexaPDF::Configuration#constantize] can now dig into hierarchical values
 
 ### Removed